Add Overflow Image Logic, Update Dropper limit calculations
An overflow line will now appear when there is a fruit that is approaching the line. It will turn red when a fruit is crossed over the line.

Adjusted the limits so the limit is now set by each fruit that is instantiated. So you can now put things closer against the wall